Chore: room 관련 API들 하나의 파일로 병합
