--- src/main/java/com/Cccccc/airbnb/domain/room/controller/CategoryController.java
@@ -1,23 +1,0 @@
-package com.Cccccc.airbnb.domain.room.controller;
-
-import com.Cccccc.airbnb.domain.room.dto.CategoryDto;
-import com.Cccccc.airbnb.domain.room.service.CategoryService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/c2")
-@RequiredArgsConstructor
-public class CategoryController {
-
-    private final CategoryService categoryService;
-
-    @GetMapping
-    public List<CategoryDto> getCategories() {
-        return categoryService.getAllCategories();
-    }
-}
