+++ src/main/java/com/Cccccc/airbnb/domain/room/controller/RoomListController.java
@@ -1,7 +1,7 @@
 package com.Cccccc.airbnb.domain.room.controller;
 
-import com.Cccccc.airbnb.domain.room.dto.RoomFilterDto;
-import com.Cccccc.airbnb.domain.room.dto.RoomListDto;
+import com.Cccccc.airbnb.domain.room.dto.request.RoomFilterRequestDto;
+import com.Cccccc.airbnb.domain.room.dto.response.RoomListResponseDto;
 import com.Cccccc.airbnb.domain.room.service.RoomListService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -12,15 +12,14 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/api/c2/room")
 @RequiredArgsConstructor
-@CrossOrigin(origins = "http://localhost:3000")
 public class RoomListController {
     private final RoomListService roomListService;
 
     @GetMapping
-    public Page<RoomListDto> getFilteredRooms(@RequestParam(defaultValue = "0") int page,
-                                              @RequestParam(defaultValue = "20") int size,
-                                              @ModelAttribute RoomFilterDto roomFilterDto) {
+    public Page<RoomListResponseDto> getFilteredRooms(@RequestParam(defaultValue = "0") int page,
+                                                      @RequestParam(defaultValue = "20") int size,
+                                                      @ModelAttribute RoomFilterRequestDto roomFilterRequestDto) {
         Pageable pageable = PageRequest.of(page, size);
-        return roomListService.getFilteredRooms(pageable, roomFilterDto);
+        return roomListService.getFilteredRooms(pageable, roomFilterRequestDto);
     }
 }
