+++ src/main/java/com/Cccccc/airbnb/domain/room/dto/response/AmenitiesResponseDto.java
@@ -1,11 +1,11 @@
-package com.Cccccc.airbnb.domain.room.dto;
+package com.Cccccc.airbnb.domain.room.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-public class Amenities {
+public class AmenitiesResponseDto {
     private String amenityName;
     private String icon;
 }
