+++ src/main/java/com/Cccccc/airbnb/domain/room/dto/response/CategoryResponseDto.java
@@ -1,11 +1,11 @@
-package com.Cccccc.airbnb.domain.room.dto;
+package com.Cccccc.airbnb.domain.room.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-public class CategoryDto {
+public class CategoryResponseDto {
     private String categoryName;
     private String image;
 }
