+++ src/main/java/com/Cccccc/airbnb/domain/room/repository/FacilityRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RoomFacilityRepository extends JpaRepository<Facility, Integer> {
+public interface FacilityRepository extends JpaRepository<Facility, Integer> {
     @Query("SELECT new com.Cccccc.airbnb.domain.room.dto.response.RoomFacilityResponseDto(f.facility_name, f.facility_image_name) " +
             "FROM Facility f JOIN RoomFacility rf ON f.id = rf.facility.id " +
             "WHERE rf.room.id = :id")
